Number now has a 'number_cast' attribute for handling alternative bases
A number can be created with a reference to a function to represent the Number as whatever base or format it originally was, when casting to a string.

Binary, hexadecimal, scientific notation numbers are now all stored as Number objects inside the calculator, and only changed back to the original representation when casting to string.